src/main/java/com/filmlib/service/impl/UserServiceImpl.java
package com.filmlib.service.impl;

import com.filmlib.entity.User;
import com.filmlib.repository.UserRepo;
import com.filmlib.service.UserService;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
public class UserServiceImpl implements UserService {
    private final UserRepo userRepo;

    public UserServiceImpl(UserRepo userRepo) {
        this.userRepo = userRepo;
    }

    @Override
    @Transactional
    public User loadUserByUsername(String username) {
        return userRepo.findByUsername(username);
    }

    @Override
    public User save(User user) {
        return userRepo.save(user);
    }

}
